Fix rounding errors in bilinearVibration
I was getting a lot of very funky rounding errors because the ranges were being
passed as integers. Now I have the function treat the passed integers as
doubles and everything's fine.

+++ src/io.h
@@ -10,9 +10,9 @@
 
 int getDistance(const int &serial_connection);
 int logarithmicVibration(const int &distance);
-int bilinearVibration(const int &distance, const int &min_distance,
-        const int &max_distance, const int &min_vibration,
-        const int &max_vibration);
+int bilinearVibration(const int &distance, const double &min_distance,
+        const double &max_distance, const double &min_vibration,
+        const double &max_vibration);
 double valueToPercent(const double &value, const double &min, const double &max);
 double percentToValue(const double &percent, const double &min, const double &max);
